Python-Api/requirements.txt
pyserial==3.4
numpy